validate/requirements.txt
jsonschema==4.22.0
PyYAML==6.0.1
